--- a/Vagrantfile
+++ b/Vagrantfile
@@ -79,14 +79,13 @@ Vagrant.configure(2) do |config|
   # Mount the gitignored puppet/modules directory, for caching
   config.vm.synced_folder "puppet/modules", "/etc/puppet/modules"
 
-  # Throw more resources at the VM. Tweak as needed
+  # Configure the VM. Tweak as needed
+  configured = -1
   config.vm.provider :virtualbox do |vb|
     # Name this virtual machine "precip"
     vb.customize ["modifyvm", :id, "--name", "precip"]
-    # IOAPIC is needed for a 64bit host
+    # IOAPIC is needed for a 64bit host for multiple cures
     vb.customize ["modifyvm", :id, "--ioapic", "on"]
-    # A single processor still outperforms larger numbers with Virtualbox
-    vb.customize ["modifyvm", :id, "--cpus", "2"]
     # Use ICH9 for performance
     vb.customize ["modifyvm", :id, "--chipset", "ich9"]
     # Deffer DNS resolution to the host for performance
@@ -95,20 +94,37 @@ Vagrant.configure(2) do |config|
     vb.customize ["modifyvm", :id, "--vram", "10"]
     # Prevent a time drift of more than a minute from the host
     vb.customize ["guestproperty", "set", :id, "/VirtualBox/GuestAdd/VBoxService/--timesync-set-threshold", 60000]
-    # Give this virtual machine 1/4th of the host RAM
+    # Give VM 1/4 system memory (if more than 2GB), otherwise leave as default
+    # Give half of cpu cores as the host (if more than 1), otherwise leave as default
+    # In our testing this produced the best results. Adapted from https://github.com/rdsubhas/vagrant-faster
     host = RbConfig::CONFIG['host_os']
+    cpus = -1
+    mem = -1
     if host =~ /darwin/
-      # sysctl returns Bytes and we need to convert to MB
-      mem = `sysctl -n hw.memsize`.to_i / 1024
+      cpus = `sysctl -n hw.ncpu`.to_i
+      mem = `sysctl -n hw.memsize`.to_i / 1024 / 1024
     elsif host =~ /linux/
-      # meminfo shows KB and we need to convert to MB
-      mem = `grep 'MemTotal' /proc/meminfo | sed -e 's/MemTotal://' -e 's/ kB//'`.to_i 
+      cpus = `nproc`.to_i
+      mem = `grep 'MemTotal' /proc/meminfo | sed -e 's/MemTotal://' -e 's/ kB//'`.to_i / 1024
     elsif host =~ /mswin|mingw|cygwin/
-      # Windows code via https://github.com/rdsubhas/vagrant-faster
-      mem = `wmic computersystem Get TotalPhysicalMemory`.split[1].to_i / 1024
+      cpus = `wmic cpu Get NumberOfCores`.split[1].to_i
+      mem = `wmic computersystem Get TotalPhysicalMemory`.split[1].to_i / 1024 / 1024
     end
-    mem = mem / 1024 / 4
-    vb.customize ["modifyvm", :id, "--memory", mem]
+    cpus = cpus / 2 if cpus > 1
+    mem = mem / 4 if mem > 2048
+    if mem > 0
+      vb.customize ["modifyvm", :id, "--memory", mem]
+      if configured < 1
+        puts "Memory set to: #{mem}"
+      end
+    end
+    if cpus > 0
+      vb.customize ["modifyvm", :id, "--cpus", cpus]
+      if configured < 1
+        puts "CPUs set to: #{cpus}"
+      end
+    end
+    configured = 1
   end
 
   # Fix the harmless "stdin: is not a tty" issue once and for all