--- a/Vagrantfile
+++ b/Vagrantfile
@@ -94,30 +94,21 @@ Vagrant.configure(2) do |config|
     vb.customize ["modifyvm", :id, "--vram", "10"]
     # Prevent a time drift of more than a minute from the host
     vb.customize ["guestproperty", "set", :id, "/VirtualBox/GuestAdd/VBoxService/--timesync-set-threshold", 60000]
-    # Give VM 1/4 system memory (if more than 2GB), otherwise leave as default
+    # Set reserved memory at 2GB, as we've never seen it actively use more than ~1.5G
+    vb.customize ["modifyvm", :id, "--memory", 2048]
     # Give half of cpu cores as the host (if more than 1), otherwise leave as default
     # In our testing this produced the best results. Adapted from https://github.com/rdsubhas/vagrant-faster
     host = RbConfig::CONFIG['host_os']
     cpus = -1
-    mem = -1
     if host =~ /darwin/
       cpus = `sysctl -n hw.ncpu`.to_i
       mem = `sysctl -n hw.memsize`.to_i / 1024 / 1024
     elsif host =~ /linux/
       cpus = `nproc`.to_i
-      mem = `grep 'MemTotal' /proc/meminfo | sed -e 's/MemTotal://' -e 's/ kB//'`.to_i / 1024
     elsif host =~ /mswin|mingw|cygwin/
       cpus = `wmic cpu Get NumberOfCores`.split[1].to_i
-      mem = `wmic computersystem Get TotalPhysicalMemory`.split[1].to_i / 1024 / 1024
     end
     cpus = cpus / 2 if cpus > 1
-    mem = mem / 4 if mem > 2048
-    if mem > 0
-      vb.customize ["modifyvm", :id, "--memory", mem]
-      if configured < 1
-        puts "Memory set to: #{mem}"
-      end
-    end
     if cpus > 0
       vb.customize ["modifyvm", :id, "--cpus", cpus]
       if configured < 1