@heathdutton 2016-10-13
@jeffgeorge do you suggest we sit a static ram limit, or set a maximum?