@jeffgeorge 2016-10-13
I think a static 2Gb of reserved RAM is fine, i've never seen it max out on RAM using 2.5G, so I think going to a proportional 1/4 of host RAM would be overkill.